+++ vector/src/main/java/im/vector/riotx/core/error/fatal.kt
@@ -16,14 +16,13 @@
 
 package im.vector.riotx.core.error
 
-import im.vector.riotx.BuildConfig
 import timber.log.Timber
 
 /**
  * throw in debug, only log in production. As this method does not always throw, next statement should be a return
  */
-fun fatalError(message: String) {
-    if (BuildConfig.DEBUG) {
+fun fatalError(message: String, failFast: Boolean) {
+    if (failFast) {
         error(message)
     } else {
         Timber.e(message)